# Add `CliDefinitionLoader` for provider module level CLI Commands