@jason810496 2025-03-01
As we introduce provider module-level CLI commands, `ProviderManager` is now initialized within `cli_parser`. Should I remove the `airflow providers lazy-loaded` command, or modify as that `ProviderManager` is only initialized once within `cli_parser`?  

cc @potiuk